do some refactoring
