+++ app/src/main/java/com/zxm/okhttp/network/OkhttpService.java
@@ -1,4 +1,4 @@
-package com.zxm.okhttp.net;
+package com.zxm.okhttp.network;
 
 import com.zxm.okhttp.entiy.BaseUserInfo;
